refactor(db): reorganize utility functions for clarity
